Shifted error handling out of dispatch
