+++ src/main.rs
@@ -9,8 +9,6 @@ use clap::Parser; // Needs to be in scope for the derive macro (::parse()) to wo
 // const MOCK_LOG_PATH: &str = "mockdb/logs.txt";
 
 fn main() {
-    backend::setup();
-
     let args = cli_parsing::Args::parse();
 
     dbg!(&args);
