+++ src/modes/mod.rs
@@ -1,5 +1,5 @@
 /// This module contains all the modes of the application.
 pub mod alter;
 pub mod print;
-pub mod quick_note;
+// pub mod quick_note;
 pub mod run_quiz;
