+++ src/modes/print.rs
@@ -5,19 +5,18 @@ use crate::backend::api;
 
 /// Prints the entries from the last `days` days.
 /// If `output` is `None`, the entries are printed to stdout. Otherwise, they are written to the file at `output`.
-pub fn print(days: u32, output: &Option<String>) -> Result<(), crate::errors::Error> {
+pub fn print(days: u32, output: &Option<String>) {
     let starting_date = chrono::Utc::now().naive_utc() - chrono::Duration::days(days as i64);
     let end_date = chrono::Utc::now().naive_utc();
 
-    let printable = printable_entries(starting_date, end_date)?;
+    let printable = printable_entries(starting_date, end_date).unwrap();
     if printable.is_empty() {
-        println!("No entries found");
-        return Ok(());
+        panic!("No entries found");
     }
 
     match output {
         Some(path) => {
-            std::fs::write(path, printable)?;
+            std::fs::write(path, printable).unwrap();
             // TODO: Add a message saying that the entries were written to the file.
             // TODO: Error handling here if this fails.
         }
@@ -25,8 +24,6 @@ pub fn print(days: u32, output: &Option<String>) -> Result<(), crate::errors::Er
             println!("{}", printable);
         }
     }
-
-    Ok(())
 }
 
 /// Returns a formatted string containing all the entries between the two dates.
